pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]

[project]
name = "scout"
readme = "README.md"
dynamic = ["version"]
authors = [
  {email = "chioke.harris@nrel.gov"},
  {name = "Chioke Harris"},
  {email = "jared.langevin@lbl.gov"},
  {name = "Jared Langevin"}
]
requires-python = ">=3.8"
license = {file = "LICENSE.md"}
dependencies = [
    "pandas >= 2.0",
    "numpy >= 1.16",
    "scipy",
    "requests",
    "numpy-financial",
    "xlsxwriter",
    "matplotlib",
    "backoff",
    "pyyaml",
    "jsonschema"
]

[tool.setuptools_scm]

[project.optional-dependencies]
dev = [
    "openpyxl",
    "flake8 == 6.0",
]